requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.1
sqlalchemy==2.0.23
psycopg[binary]==3.1.13
python-dotenv==1.0.0
bcrypt==4.1.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
